refactor: created util folder, moving funcs from reducer to utils, renaming playerHands to players, setting up initial players state
