+++ package.json
@@ -80,6 +80,7 @@
     "cross-env": "^3.1.4",
     "dateformat": "^2.0.0",
     "eslint": "^3.19.0",
+    "eslint-config-fullstack": "^3.0.0",
     "eslint-config-standard": "^10.2.1",
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-node": "^4.2.2",
